+++ frontend/src/libs/path.test.js
@@ -5,8 +5,20 @@ describe("suite", () => {
     it("parts", () => {
         expect(new Path("/foo/bar").parts).toStrictEqual(["/", "foo", "bar"]);
         expect(new Path("//foo/bar").parts).toStrictEqual(["/", "foo", "bar"]);
-        expect(new Path("/./foo/bar").parts).toStrictEqual(["/", "foo", "bar"]);
-        expect(new Path("./foo/bar").parts).toStrictEqual(["foo", "bar"]);
-        expect(new Path("./foo/./bar").parts).toStrictEqual(["foo", "bar"]);
+        expect(new Path("/./foo/bar").parts).toStrictEqual(["/", ".", "foo", "bar"]);
+        expect(new Path("./foo/bar").parts).toStrictEqual([".", "foo", "bar"]);
+        expect(new Path("./foo/./bar").parts).toStrictEqual([".", "foo", ".", "bar"]);
+    })
+    it("resolve", () => {
+        expect(new Path("/foo/bar").resolve().parts).toStrictEqual(["/", "foo", "bar"]);
+        expect(new Path("//foo/bar").resolve().parts).toStrictEqual(["/", "foo", "bar"]);
+        expect(new Path("/./foo/bar").resolve().parts).toStrictEqual(["/", "foo", "bar"]);
+        expect(new Path("./foo/bar").resolve().parts).toStrictEqual(["foo", "bar"]);
+        expect(new Path("./foo/./bar").resolve().parts).toStrictEqual(["foo", "bar"]);
+
+        expect(new Path("/foo/bar/..").resolve().parts).toStrictEqual(["/", "foo"]);
+        expect(new Path("/foo/bar/../").resolve().parts).toStrictEqual(["/", "foo"]);
+        expect(new Path("/foo/bar/..//").resolve().parts).toStrictEqual(["/", "foo"]);
+        expect(new Path("/foo/bar/../../foo").resolve().parts).toStrictEqual(["/", "foo"]);
     })
 });